# docs/swagger.yaml
basePath: /api/
definitions:
  api.Author:
    properties:
      id:
        description: Id of the author
        type: integer
      name:
        description: Name of the author
        type: string
      url:
        description: external url of the author
        type: string
    type: object
  api.AuthorResponse:
    properties:
      filter:
        description: filter/query information
        items:
          $ref: '#/definitions/api.Filter'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: list pagination
      result:
        description: authors
        items:
          $ref: '#/definitions/api.ListAuthor'
        type: array
    type: object
  api.Book:
    properties:
      authors:
        description: list of authors
        items:
          $ref: '#/definitions/api.Author'
        type: array
      blurb:
        description: blurb if the book
        type: string
      coverColor:
        description: If has cover image, itas the avarage cover color
        type: string
      coverType:
        description: cover filetype (if present cover image)
        type: string
      edited:
        description: True if the metadata at least once has be saved (it will not
          overwriten by any automatic process)
        type: integer
      file:
        description: file of the book
        type: string
      fileType:
        description: 'type of the ebook: epub or mobipocket (prc, azw)'
        type: string
      hasCover:
        description: Book has cover image
        type: boolean
      id:
        description: id of the book
        type: integer
      series:
        description: list of series
        items:
          $ref: '#/definitions/api.Series'
        type: array
      tags:
        description: list of tags
        items:
          $ref: '#/definitions/api.Tag'
        type: array
      title:
        description: Title of the book (if present)
        type: string
      url:
        description: external Url of the book
        type: string
    type: object
  api.BookResponse:
    properties:
      display:
        allOf:
        - $ref: '#/definitions/api.ExtraDisplay'
        description: Extra display information
      filter:
        description: filter/query information
        items:
          $ref: '#/definitions/api.Filter'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: list pagination
      result:
        description: list opf the books
        items:
          $ref: '#/definitions/api.Book'
        type: array
    type: object
  api.ExtraDisplay:
    properties:
      data:
        description: data for the filtering
        type: integer
      type:
        description: type, description of extra information
        type: string
    type: object
  api.Filter:
    properties:
      type:
        description: type of the filter
        type: string
      value:
        description: value of the filter
        type: string
    type: object
  api.ListAuthor:
    properties:
      books:
        description: book count of the author
        type: integer
      id:
        description: Id of the author
        type: integer
      name:
        description: Name of the author
        type: string
      url:
        description: external url of the author
        type: string
    type: object
  api.ListSeries:
    properties:
      books:
        description: book count in this series
        type: integer
      id:
        description: Id of the series
        type: integer
      name:
        description: Name (title) of the series
        type: string
      seqno:
        description: Book order in the series
        type: integer
      url:
        description: external Url of the series
        type: string
    type: object
  api.Pagination:
    properties:
      base:
        description: bas url for reloding
        type: string
      pages:
        description: number of tha pages
        type: integer
      perPage:
        description: results per page
        type: integer
      qty:
        description: number of the all results
        type: integer
      selectedPage:
        description: selected, actual page number
        type: integer
    type: object
  api.ScraperAuthor:
    properties:
      name:
        description: name of the author
        type: string
      url:
        description: external Url
        type: string
    type: object
  api.ScraperResult:
    properties:
      authors:
        description: list of authors
        items:
          $ref: '#/definitions/api.ScraperAuthor'
        type: array
      blurb:
        description: blurb of the book
        type: string
      series:
        description: list of the series
        items:
          $ref: '#/definitions/api.ScraperSeries'
        type: array
      tags:
        description: list of the tags
        items:
          type: string
        type: array
      title:
        description: title of the book
        type: string
    type: object
  api.ScraperSeries:
    properties:
      name:
        description: name of the series
        type: string
      seqno:
        description: book's position in the series
        type: string
      url:
        description: external Url
        type: string
    type: object
  api.Series:
    properties:
      id:
        description: Id of the series
        type: integer
      name:
        description: Name (title) of the series
        type: string
      seqno:
        description: Book order in the series
        type: integer
      url:
        description: external Url of the series
        type: string
    type: object
  api.SeriesResponse:
    properties:
      filter:
        description: filter/query information
        items:
          $ref: '#/definitions/api.Filter'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: list pagination
      result:
        description: Series
        items:
          $ref: '#/definitions/api.ListSeries'
        type: array
    type: object
  api.SuggestionData:
    properties:
      id:
        description: id of the data
        type: integer
      name:
        description: Text of the data
        type: string
      url:
        description: url of the data
        type: string
    type: object
  api.SuggestionItem:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/api.SuggestionData'
        description: data for processing
      value:
        description: display text
        type: string
    type: object
  api.Suggestions:
    properties:
      query:
        description: query string
        type: string
      suggestions:
        description: result list
        items:
          $ref: '#/definitions/api.SuggestionItem'
        type: array
    type: object
  api.Tag:
    properties:
      color:
        description: Color of the tag
        type: string
      id:
        description: id fo the tag
        type: integer
      name:
        description: name of the tag
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8888
info:
  contact: {}
  description: Query/edit eBook metadata
  title: eBook library API
  version: "1.0"
paths:
  /api/scraper/:
    get:
      description: The endpont give a processed url andpoint as a book
      parameters:
      - description: The url need to scrape
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ScraperResult'
      summary: Scrape and process an url
  /author:
    get:
      description: The endpont give back a list of authors with pagging, filtering
        information
      parameters:
      - description: page number, start with 1, default is 1
        in: query
        name: page
        type: integer
      - description: A query string used to filter authors based on their names.
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AuthorResponse'
      summary: Query for Authors metadata
      tags:
      - author
  /book:
    get:
      description: The endpont give back a list of book metadata with pagging, filtering
        information
      parameters:
      - description: page number, start with 1, default is 1
        in: query
        name: page
        type: integer
      - description: author ID used to filter books based on authors
        in: query
        name: ai
        type: integer
      - description: series ID used to filter books based on series
        in: query
        name: si
        type: integer
      - description: tag ID used to filter books based on tags
        in: query
        name: ti
        type: integer
      - description: A query string used to filter books based on their description,
          title, author, series, and tags.
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BookResponse'
      summary: Query for books metadata
      tags:
      - book
    post:
      consumes:
      - application/json
      description: Uopdate a book's metadata
      parameters:
      - description: Book to update
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/api.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Update a book
      tags:
      - book
  /book/{bookId}:
    get:
      description: The endpont give back the given book netadata
      parameters:
      - description: ID of the book
        in: path
        name: bookId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Book'
        "404":
          description: Book not found
          schema:
            type: string
      summary: Get a book metadata
      tags:
      - book
  /export:
    get:
      description: The endpont give back file for every ebook in the database
      produces:
      - application/json
      responses:
        "200":
          description: JSON file containing book list
          schema:
            type: file
      summary: create a json file for all books in the system
      tags:
      - book
  /query/author:
    get:
      description: The endpont give back a list of authors for the given query string
      parameters:
      - description: A query string used to filter authors based on their names.
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Suggestions'
      summary: Query Authors for autocomplete
      tags:
      - author
  /query/tags:
    get:
      description: The endpont give back a list of tags for the given query string
      parameters:
      - description: A query string used to filter tags based on their title.
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Suggestions'
      summary: Query Tags for autocomplete
      tags:
      - tags
  /series:
    get:
      description: The endpont give back a list of series with pagging, filtering
        information
      parameters:
      - description: page number, start with 1, default is 1
        in: query
        name: page
        type: integer
      - description: A query string used to filter series based on their title/name.
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SeriesResponse'
      summary: Query for Series metadata
      tags:
      - series
securityDefinitions:
  none:
    type: basic
swagger: "2.0"